requirements.txt
streamlit==1.17.0
torch==2.0.1
torchvision==0.15.1
Pillow==10.0.0
numpy==1.25.2
opencv-python-headless==4.7.0.72
matplotlib==3.8.0
